+++ src/ios/AppDelegate+BranchSdk.m
@@ -43,6 +43,13 @@
   return YES;
 }
 
+- (BOOL)application:(UIApplication*)application didFinishLaunchingWithOptions:(NSDictionary*)launchOptions
+{
+    // --- ADD THIS ---
+    [[Branch getInstance] enableLogging];
+    return YES;
+}
+
 // Respond to Push Notifications
 - (void)application:(UIApplication *)application didReceiveRemoteNotification:(NSDictionary *)userInfo {
   @try {